Core - Fix/suppress some compiler warnings.
size_t -> int conversion warnings

Still lots to resolve, some properties like frameidentifier could be changed to Int64

+++ CefSharp.Core/Internals/CefBrowserWrapper.cpp
@@ -210,7 +210,7 @@ IFrame^ CefBrowserWrapper::GetFrame(String^ name)
 int CefBrowserWrapper::GetFrameCount()
 {
     ThrowIfDisposed();
-    return _browser->GetFrameCount();
+    return static_cast<int>(_browser->GetFrameCount());
 }
 
 ///
@@ -223,7 +223,7 @@ List<Int64>^ CefBrowserWrapper::GetFrameIdentifiers()
 
     std::vector<Int64> identifiers;
     _browser->GetFrameIdentifiers(identifiers);
-    List<Int64>^ results = gcnew List<Int64>(identifiers.size());
+    List<Int64>^ results = gcnew List<Int64>(static_cast<int>(identifiers.size()));
     for (UINT i = 0; i < identifiers.size(); i++)
     {
         results->Add(identifiers[i]);
